Better buffer management with remote timestamps
The data format now includes a local timestamp (remote timestamp to the
other side). By checking the remote timestamp and the arrival timestamp,
a client can't determine the ping, but *can* determine the effective
variance in ping, and a buffer has to be enough to subsume ping
variance. The ideal buffer size is now based on the max of ping and this
variance.

This is a particularly useful source once we add video, because
keyframes take longer to send than delta frames.